+++ xds/src/main/java/com/linecorp/armeria/xds/ClusterRoot.java
@@ -18,7 +18,7 @@ package com.linecorp.armeria.xds;
 
 import com.linecorp.armeria.common.annotation.Nullable;
 import com.linecorp.armeria.common.annotation.UnstableApi;
-import com.linecorp.armeria.xds.SnapshotStream.Subscription;
+import com.linecorp.armeria.xds.stream.Subscription;
 
 import io.envoyproxy.envoy.config.cluster.v3.Cluster;
