+++ xds/src/main/java/com/linecorp/armeria/xds/ListenerStream.java
@@ -19,6 +19,9 @@ package com.linecorp.armeria.xds;
 import static com.linecorp.armeria.xds.XdsType.LISTENER;
 
 import com.linecorp.armeria.common.annotation.Nullable;
+import com.linecorp.armeria.xds.stream.RefCountedStream;
+import com.linecorp.armeria.xds.stream.SnapshotStream;
+import com.linecorp.armeria.xds.stream.Subscription;
 
 import io.envoyproxy.envoy.config.core.v3.ConfigSource;
 import io.envoyproxy.envoy.config.route.v3.RouteConfiguration;
