+++ xds/src/main/java/com/linecorp/armeria/xds/RawBufferTransportSocketFactory.java
@@ -21,6 +21,7 @@ import java.util.List;
 import com.google.common.collect.ImmutableList;
 
 import com.linecorp.armeria.common.annotation.Nullable;
+import com.linecorp.armeria.xds.stream.SnapshotStream;
 
 import io.envoyproxy.envoy.config.core.v3.ConfigSource;
 import io.envoyproxy.envoy.config.core.v3.TransportSocket;
