+++ xds/src/main/java/com/linecorp/armeria/xds/UpstreamTlsTransportSocketFactory.java
@@ -22,6 +22,7 @@ import java.util.Optional;
 import com.google.common.collect.ImmutableList;
 
 import com.linecorp.armeria.common.annotation.Nullable;
+import com.linecorp.armeria.xds.stream.SnapshotStream;
 
 import io.envoyproxy.envoy.config.core.v3.ConfigSource;
 import io.envoyproxy.envoy.config.core.v3.TransportSocket;
