+++ xds/src/main/java/com/linecorp/armeria/xds/stream/CombineLatest3Stream.java
@@ -14,35 +14,43 @@
  * under the License.
  */
 
-package com.linecorp.armeria.xds;
-
-import java.util.function.BiFunction;
+package com.linecorp.armeria.xds.stream;
 
 import com.linecorp.armeria.common.annotation.Nullable;
+import com.linecorp.armeria.xds.SnapshotWatcher;
 
-final class CombineLatest2Stream<A, B, O> extends RefCountedStream<O> {
+final class CombineLatest3Stream<A, B, C, O> extends RefCountedStream<O> {
 
     private final SnapshotStream<A> streamA;
     private final SnapshotStream<B> streamB;
-    private final BiFunction<? super A, ? super B, ? extends O> combiner;
+    private final SnapshotStream<C> streamC;
+    private final TriFunction<? super A, ? super B, ? super C, ? extends O> combiner;
 
     @Nullable
     private Subscription subA;
     @Nullable
     private Subscription subB;
+    @Nullable
+    private Subscription subC;
 
     private boolean aReady;
     private boolean bReady;
+    private boolean cReady;
+
     @Nullable
     private A latestA;
     @Nullable
     private B latestB;
+    @Nullable
+    private C latestC;
 
-    CombineLatest2Stream(SnapshotStream<A> streamA,
+    CombineLatest3Stream(SnapshotStream<A> streamA,
                          SnapshotStream<B> streamB,
-                         BiFunction<? super A, ? super B, ? extends O> combiner) {
+                         SnapshotStream<C> streamC,
+                         TriFunction<? super A, ? super B, ? super C, ? extends O> combiner) {
         this.streamA = streamA;
         this.streamB = streamB;
+        this.streamC = streamC;
         this.combiner = combiner;
     }
 
@@ -50,6 +58,7 @@ final class CombineLatest2Stream<A, B, O> extends RefCountedStream<O> {
     protected Subscription onStart(SnapshotWatcher<O> watcher) {
         subA = streamA.subscribe(this::onA);
         subB = streamB.subscribe(this::onB);
+        subC = streamC.subscribe(this::onC);
 
         return () -> {
             if (subA != null) {
@@ -58,8 +67,12 @@ final class CombineLatest2Stream<A, B, O> extends RefCountedStream<O> {
             if (subB != null) {
                 subB.close();
             }
+            if (subC != null) {
+                subC.close();
+            }
             subA = null;
             subB = null;
+            subC = null;
         };
     }
 
@@ -68,7 +81,6 @@ final class CombineLatest2Stream<A, B, O> extends RefCountedStream<O> {
             emit(null, t);
             return;
         }
-
         latestA = v;
         aReady = true;
         maybeEmit();
@@ -79,18 +91,30 @@ final class CombineLatest2Stream<A, B, O> extends RefCountedStream<O> {
             emit(null, t);
             return;
         }
-
         latestB = v;
         bReady = true;
         maybeEmit();
     }
 
-    private void maybeEmit() {
-        if (!aReady || !bReady) {
+    private void onC(@Nullable C v, @Nullable Throwable t) {
+        if (v == null) {
+            emit(null, t);
             return;
         }
+        latestC = v;
+        cReady = true;
+        maybeEmit();
+    }
+
+    private void maybeEmit() {
+        if (!aReady || !bReady || !cReady) {
+            return;
+        }
+        assert latestA != null;
+        assert latestB != null;
+        assert latestC != null;
         try {
-            emit(combiner.apply(latestA, latestB), null);
+            emit(combiner.apply(latestA, latestB, latestC), null);
         } catch (Throwable t) {
             emit(null, t);
         }
