+++ xds/src/test/java/com/linecorp/armeria/xds/stream/RefCountedStreamTest.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.xds;
+package com.linecorp.armeria.xds.stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -28,8 +28,7 @@ import java.util.function.Supplier;
 
 import org.junit.jupiter.api.Test;
 
-import com.linecorp.armeria.common.annotation.Nullable;
-import com.linecorp.armeria.xds.SnapshotStream.Subscription;
+import com.linecorp.armeria.xds.SnapshotWatcher;
 
 class RefCountedStreamTest {
 
@@ -322,11 +321,6 @@ class RefCountedStreamTest {
         @Override
         protected void onStop() {
             onStopCallback.run();
-        }
-
-        @Override
-        protected void emit(@Nullable String value, @Nullable Throwable error) {
-            super.emit(value, error);
         }
     }
 }
